refactor: implement ui, storage, parser, commands classes
