+++ data/tasks.txt
@@ -1,4 +1,6 @@
 T | 1 | tests13
-T | 0 | tesfr1
-D | 0 | jksadjsla | 20/02/2000 00:00
+T | 1 | tesfr1
 E | 0 | aidjskdjsa | 20/02/2000 00:00 | 21/02/2000 12:09
+T | 0 | eowfejio
+D | 0 | aksokda | 22/12/2200 00:00
+E | 0 | asdas | 12/02/1111 00:29 | 12/02/2313 00:00
