+++ src/main/java/command/EventCommand.java
@@ -1,0 +1,39 @@
+package command;
+
+import exceptions.CommandException;
+import exceptions.InvalidCommandException;
+import storage.Storage;
+import task.EventTask;
+import task.TaskList;
+import util.Helper;
+
+import java.time.LocalDateTime;
+
+public class EventCommand implements Command {
+    private final String args;
+
+    public EventCommand(String args) {
+        this.args = args;
+    }
+
+    /**
+     * Handles the event command to add a new event task.
+     */
+    @Override
+    public String execute(TaskList taskList, Storage storage) throws CommandException {
+        String[] parts = args.split("\\s+/from\\s+|\\s+/to\\s+");
+        if (parts.length < 3 || parts[0].isBlank() || parts[1].isBlank() || parts[2].isBlank()) {
+            throw new InvalidCommandException("event <description> /from <dd/MM/yyyy> <HH:mm> /to <dd/MM/yyyy> <HH:mm>");
+        }
+        LocalDateTime startTime = Helper.parseDateTime(parts[1].trim());
+        LocalDateTime endTime = Helper.parseDateTime(parts[2].trim());
+        if (startTime == null || endTime == null) {
+            throw new InvalidCommandException("event <description> /from <dd/MM/yyyy> <HH:mm> /to <dd/MM/yyyy> <HH:mm>");
+        }
+
+        String response = taskList.addTask(new EventTask(parts[0].trim(), startTime, endTime));
+        storage.saveTasks(taskList);
+
+        return response;
+    }
+}
