+++ src/main/java/task/TaskList.java
@@ -12,161 +12,67 @@ import java.nio.file.Paths;
 import java.util.ArrayList;
 
 public class TaskList {
-    private static final String DATA_FILE_PATH = "./data/tasks.txt";
     private ArrayList<Task> tasks;
 
     public TaskList() {
         this.tasks = new ArrayList<>();
-        loadTask();
     }
 
-    private void saveTasks() {
-        Path filePath = Paths.get(DATA_FILE_PATH);
-
-        try {
-            Files.createDirectories(filePath.getParent());
-
-            try (BufferedWriter writer = Files.newBufferedWriter(filePath)) {
-                for (Task task : tasks) {
-                    writer.write(task.toFileFormat());
-                    writer.newLine();
-                }
-            }
-        } catch (IOException e) {
-            System.out.println("\t Error saving tasks: " + e.getMessage());
-        }
+    public ArrayList<Task> getTasks() {
+        return tasks;
     }
 
-    private void loadTask() {
-        Path filePath = Paths.get(DATA_FILE_PATH);
-        try {
-            Files.createDirectories(filePath.getParent());
-        } catch (IOException e) {
-            System.out.println("\t Warning: Could not create data directory");
-            return;
-        }
-
-        if (!Files.exists(filePath)) {
-            System.out.println("\t No existing data file found. Starting with empty task list.");
-            return;
-        }
-
-        try (BufferedReader reader = Files.newBufferedReader(filePath)) {
-            String line;
-            int lineNumber = 0;
-            int corruptedLines = 0;
-
-            while ((line = reader.readLine()) != null) {
-                lineNumber++;
-                try {
-                    Task task = parseTaskFromLine(line);
-                    if (task != null) {
-                        tasks.add(task);
-                    }
-                } catch (IllegalArgumentException e) {
-                    System.out.println("\t Warning: Skipping corrupted line " + lineNumber + ": " +
-                                        line + " (" + e.getMessage() + ")");
-                    corruptedLines++;
-                }
-            }
-
-            System.out.println("\t Loaded " + tasks.size() + " tasks from data file.");
-            if (corruptedLines > 0) {
-                System.out.println("\t Skipped " + corruptedLines + " corrupted lines. File will be cleaned up.");
-                saveTasks();
-            }
-
-        } catch (IOException e) {
-            System.out.println("\t Error reading data file: " + e.getMessage());
-        }
-    }
-
-    private Task parseTaskFromLine(String line) {
-        String[] parts = line.split(" \\| ");
-        if (parts.length < 3) {
-            throw new IllegalArgumentException("Invalid format");
-        }
-
-        String type = parts[0];
-        boolean isDone = parts[1].equals("1");
-        String description = parts[2];
-
-        Task task;
-        switch (type) {
-        case "T":
-            task = new ToDoTask(description);
-            break;
-        case "D":
-            if (parts.length < 4) {
-                throw new IllegalArgumentException("Missing deadline");
-            }
-            task = new DeadlineTask(description, Helper.parseDateTime(parts[3]));
-            break;
-        case "E":
-            if (parts.length < 5) {
-                throw new IllegalArgumentException("Missing event time");
-            }
-            task = new EventTask(description, Helper.parseDateTime(parts[3]), Helper.parseDateTime(parts[4]));
-            break;
-        default:
-            throw new IllegalArgumentException("Unknown task type: " + type);
-        }
-
-        if (isDone) {
-            task.markAsDone();
-        }
-
-        return task;
-    }
-
-    public void addTask(Task task) {
+    public String addTask(Task task) {
         this.tasks.add(task);
-        System.out.println("\t I've scribbled down your little task:");
-        System.out.println("\t  " + task.getDescription());
-        System.out.println("\t Now, do try to keep up, won't you?");
-        System.out.println("\t You have " + this.tasks.size() + " tasks left.");
-        saveTasks();
+        return "\t I've scribbled down your little task:\n" +
+                "\t  " + task.getDescription() + "\n" +
+                "\t Now, do try to keep up, won't you?\n" +
+                "\t You have " + this.tasks.size() + " tasks left.\n";
     }
 
-    public void markTask(int i) {
+    public String markTask(int i) {
         Task task = this.tasks.get(i - 1);
         task.markAsDone();
-        System.out.println("\t Behold! I've declared this paltry task complete.");
-        System.out.println("\t  " + task.getDescription());
-        System.out.println("\t Don't get cocky. You still have a long way to go.");
-        saveTasks();
+        return "\t Behold! I've declared this paltry task complete.\n" +
+                "\t  " + task.getDescription() + "\n" +
+                "\t Don't get cocky. You still have a long way to go.\n";
     }
 
-    public void unmarkTask(int i) {
+    public String unmarkTask(int i) {
         Task task = this.tasks.get(i - 1);
         task.unmark();
-        System.out.println("\t You're toying with me! I've marked this back as incomplete.");
-        System.out.println("\t  " + task.getDescription());
-        System.out.println("\t Don't think for a second I'll forget this betrayal.");
-        saveTasks();
+        return "\t You're toying with me! I've marked this back as incomplete.\n" +
+                "\t  " + task.getDescription() + "\n" +
+                "\t Don't think for a second I'll forget this betrayal.\n";
     }
 
-    public void listTask() {
+    public String listTask() {
         if (this.tasks.isEmpty()) {
-            System.out.println("\t No tasks? How utterly dreadful!");
-            return;
+            return "\t No tasks? How utterly dreadful!\n";
         }
 
-        System.out.println("\t These, my dear simpleton, are the items on your agenda.");
+        StringBuilder sb = new StringBuilder();
+        sb.append("\t These, my dear simpleton, are the items on your agenda.\n");
+
         int i = 1;
         for (Task task : tasks) {
-            System.out.println("\t  " + i + ". " + task.getDescription());
+            sb.append("\t  ").append(i).append(". ").append(task.getDescription()).append("\n");
             i++;
         }
-        System.out.println("\t Failure is not an option.");
+        sb.append("\t Failure is not an option.\n");
+
+        return sb.toString();
     }
 
-    public void deleteTask(int i) {
+    public String deleteTask(int i) {
         Task task = this.tasks.remove(i - 1);
-        System.out.println("\t Poof! Begone with you, you insignificant little undertaking!");
-        System.out.println("\t  " + task.getDescription());
-        System.out.println("\t Don't get cocky. You still have a long way to go.");
-        System.out.println("\t You have " + this.tasks.size() + " tasks left.");
-        saveTasks();
+        return "\t Poof! Begone with you, you insignificant little undertaking!\n" +
+                "\t  " + task.getDescription() + "\n" +
+                "\t Don't get cocky. You still have a long way to go.\n" +
+                "\t You have " + this.tasks.size() + " tasks left.\n";
+    }
+
+    public int size() {
+        return tasks.size();
     }
 }
